trying to debug delete btns

+++ public/js/delete.js
@@ -2,14 +2,17 @@ document.querySelectorAll(".deleteBtn").forEach((button) => {
   button.addEventListener("click", async (event) => {
     try {
       const blogId = event.target.dataset.blog_id;
+      if (!blogId) {
+        console.log("Blog ID not found!");
+        return;
+      }
       const response = await fetch(`/api/blogs/${blogId}`, {
         method: "DELETE",
       });
       if (response.ok) {
         event.target.closest(".card").remove();
       } else {
-        console.log(blogId);
-        throw new Error("Failed to delete book");
+        console.log("Failed to delete blog post!");
       }
     } catch (err) {
       console.error(err);
